native/posix arch: Request 64bit time_t
Let's request the 64 bit time_t API implementation from glibc.
https://sourceware.org/glibc/wiki/Y2038ProofnessDesign
by explicitly defining _TIME_BITS=64

_TIME_BITS support was added in glibc 2.34
but it should, hopefully, be safe to set it even if glibc
does not support it as it should just be ignored.

Note that as this changes the time_t definition, linking to
other precompiled libraries whose ABI uses a time_t of 32bits
would cause trouble.
For those, either recompile that libray with the same
_TIME_BITS=64 or set CONFIG_ARCH_POSIX_64BIT_TIME_T=n.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>