usb: device_next: support BOS descriptor with vendor request code
Platform capability descriptors such as MSOSv2 or WebUSB BOS have a
vendor request code that is used by the host to perform vendor-specific
requests. The device stack can use the vendor request code to call the
vendor request callbacks when it receives a request of type Vendor,
recipient Device, and bRequest value equal to the vendor request code
from the BOS descriptor node.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>