Bluetooth: HFP_HF: Avoid at status overwrote by sending
There is an issue that the sending of AT+command will
overwrite the at status when the at is handling
receiving.

Add a flag `BT_HFP_HF_FLAG_RX_ONGOING` to flag the
receiving is working.

When the flag is set, suspend the TX sending.

Add a worker to restart the tx sending.

After the flag `BT_HFP_HF_FLAG_RX_ONGOING` is cleared,
submit the TX worker to restart the tx sending.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>