drivers: nxp_s32_canxl: correct the calculating initial timing
Add a function that calculates the data phase bit timing so that
the prescaler of the data phase bit timing must be equal this one
of the nominal bit timing. Due to the hardware CANXL use the same
bit time quanta prescaler for the nominal bit timing and the
data phase bit timing.

Update the maximum values of timing that is allowed by hardware.

Signed-off-by: Cong Nguyen Huu <cong.nguyenhuu@nxp.com>