drivers: audio: dmic_nrfx_pdm: Reclassify log in PDM read when no data is ready
Currently the dmic_nrfx_pdm read implementation treats a timeout as an error, in that a logging error message is produced when no PDM data is available. However, for non-(or minimally-)blocking applications this is normal behavior. Classified this way, the logger is flooded with error messages unnecessarily, unless the log level is changed for the dmic module, which is not desirable.

This modification simply changes the log-level to debug so the application user can then decide whether or not a failed read needs to produce an error message.

Signed-off-by: Anthony Wertz <awertz@pm.me>