samples: usb: add sample for the new USB DFU implementation
The sample defines an image that uses RAMdisk and also enables the USB
DFU flash backend by default for the "slot-1" flash partition.

Assuming the user runs this sample from the slot-0 partition, uploading
to flash should work fine for evaluation purposes. There is no support
for MCUboot in this sample.

Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>