boards: opta: enable ADC, RS485 and ethernet reorganization
This set of changes enables the RS485 hardware connected to usart3 and
provides some overlays that allow for easily running the modbus rtu_client
and rtu_server samples on Opta.

A valid device tree configuration is provided for the ADCs of the 8 input
channels and the sample adc_dt works out of the box. Obviously this is not
the only possible configuration but it provides a good template for
further customization without the need to lookup the ADC GPIOs and
connections in the schematics.

Also, the `status = "okay"` was removed from ethernet configuration in the
device tree to avoid always committing 16KB of RAM for the network
buffers. As shown in the new overlay for the telnet sample it is
sufficient to add it back to get a fully working network connection.

Signed-off-by: Federico Di Gregorio <fog@dndg.it>