drivers: clock_mgmt: implement LPC55Sxx specific PLL drivers
Implement driver for the LPC55Sxx PLL IP. These SOCs have two PLLs, one
of which includes a spread spectrum generator that permits better
precision when setting output clock rates, and can be used to reduce EMI
in spread spectrum mode. These PLLs are specific to the LPC55Sxx series,
so the drivers and compatibles are named accordingly.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>