drivers: video: improve video formats documentation
Introduce a new text documentation of the video formats, describing
every bit, gives the position of the most significant bit, outline how
it is cut in bytes, and show the pixel segmentation.

Also introduce a VIDEO_PIX_FMT_..._BPP macro for each format,
helping with computing the fmt->pitch field from the drivers.

Signed-off-by: Josuah Demangeon <me@josuah.net>