usb: device_next: cdc_acm: Prevent polling for buffer in TX
In case the host doesn't pull the new data from the endpoint, the work task
would schedule itself again delayed (at the max. priority). When there is
no terminal program or active application reading the endpoint this
results in a constant polling of the endpoint burning up to 5% of the
CPU cycles.

By using a atomic flag for tx busy, the polling is solved and changed into
a postponed execution of the next work task which saves up to 5% of
CPU cycles and allows a better real-time behavior for other tasks.

Secondly, under the current structure of usbd_cdc_acm, there is a
possibility that TX data is enqueued to UDC driver before the previous
transaction is completed, hence causing dropped data. By using a
atomic flag for tx busy, no data is dropped and will only be enqueued
when there is room to forward the data.

Signed-off-by: Vincent van der Locht <vincent@synchronicit.nl>