soc: nordic: nrf53: SOC_NRF53_CPUNET_ENABLE should not depend on !BT
This causes issues as we still want the net core to be set
up on initialisation if CONFIG_BT is enabled in many cases.

The previous changes in
https://github.com/zephyrproject-rtos/zephyr/pull/74304
assumed that because this is also handled in
`bt_hci_transport_setup` that it shouldn't be done on
initialisation too.

Also, I think that
`bt_hci_transport_setup` should not be even controlling the
network core state like it is currently and should be
looked at being changed in a future PR.

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>